+++ src/app/shared/models/rumble.d.ts
@@ -24,7 +24,8 @@ export type Attribute =
   | "near"
   | "Half Stats"
   | "Shield"
-  | "Confusion";
+  | "Confusion"
+  | "Counter";
 export type TargetingPriority = "highest" | "lowest";
 export type TargetElement = Targets | Classes | Colors;
 export type Targets = "self" | "crew" | "enemies";
@@ -47,12 +48,12 @@ export type EffectEnum =
   | "boon"
   | "penalty"
   | "cleanse";
-export type AttackEffectType = "atk" | "time" | "cut" | "fixed" | "atkbase";
+export type AttackEffectType = "atk" | "time" | "cut" | "fixed" | "atkbase" | "random";
 export type RechargeEffectType =  "Special CT" | "RCV" | "percentage" | "fixed";
 export type Direction = "forward" | "radial" | "sideways";
 export type Size = "large" | "small" | "medium";
 export type ConditionComparator = "above" | "below" | "remaining" | "first" | "after" | "more" | "less";
-export type ConditionType = "stat" | "time" | "crew" | "enemies" | "trigger" | "character" | "defeat" | "dmgreceived" | "special" | "action" | "attack" | "debuff" | "damage" | "hitreceived" | "dbfreceived";
+export type ConditionType = "stat" | "time" | "crew" | "enemies" | "trigger" | "character" | "defeat" | "dmgreceived" | "special" | "action" | "attack" | "debuff" | "damage" | "hitreceived" | "dbfreceived" | "dmgdealt";
 export type TeamType  = "crew" | "enemy";
 export type Pattern = AttackPattern | HealPattern;
 export type Action = "attack" | "heal";
@@ -95,6 +96,7 @@ export interface Effect {
   effect: EffectEnum;
   targeting: Targeting;
   amount?: number;
+  amountrange?: number[];
   level?: number;
   range?: Range;
   condition?: Condition;