+++ src/app/shared/pipes/effect.pipe.ts
@@ -73,6 +73,14 @@ export class EffectPipe implements PipeTransform {
             }
             e += 'base ATK in damage';
             break;
+          case 'random':
+            if (effect.amountrange == null || effect.amountrange.length != 2) {
+              console.warn('expected amountrange to have 2 entries', effect);
+              e += 'Randomly deals fixed damage';
+            } else {
+              e += `Randomly deals between ${effect.amountrange[0]}-${effect.amountrange[1]} fixed damage`;
+            }
+            break;
           default:
             e += 'TODO:  ' + JSON.stringify(effect);
         }
@@ -105,7 +113,26 @@ export class EffectPipe implements PipeTransform {
         }
         break;
       case 'boon':
-        e += ('chance' in effect ? effect.chance + '% chance to ' : '') + ('Provoke' === this.arrayToString(effect.attributes) ? 'Provoke enemies' : 'reduce ' + this.arrayToString(effect.attributes));
+        const boonPrefix = 'chance' in effect ? effect.chance + '% chance to ' : '';
+        const boonEffects = this.arrayToString(effect.attributes);
+        let boonMessage = '';
+        switch (boonEffects) {
+          case 'Provoke':
+          case 'Counter':
+          case 'Revive':
+          case 'Haste':
+            boonMessage = `apply ${boonEffects}`;
+            if (effect.amount != null && boonEffects === 'Counter') {
+              boonMessage += ` (${effect.amount}x ATK)`;
+            }
+            if (effect.amount != null && boonEffects === 'Revive') {
+              boonMessage += ` (${effect.amount}% HP)`;
+            }
+            break;
+          default:
+            boonMessage = `reduce ${boonEffects}`;
+        }
+        e += boonPrefix.concat(boonMessage);
         break;
       case 'penalty':
         const tmpStr = this.arrayToString(effect.attributes);
@@ -142,8 +169,7 @@ export class EffectPipe implements PipeTransform {
       e += ` ${effect.repeat} times`;
     }
 
-    e += '.';
-    return e;
+    return `${this.capitalizeFirst(e)}.`;
   }
 
   arrayToString(array: any): string {
@@ -175,5 +201,9 @@ export class EffectPipe implements PipeTransform {
     }
 
     return to + ('stat' in target ? ' with the ' + target.priority + ' ' + target.stat : '');
+  }
+
+  capitalizeFirst(str: string): string {
+    return str && str.charAt(0).toUpperCase().concat(str.slice(1));
   }
 }