+++ src/CHANGELOG.md
@@ -1,5 +1,8 @@
 ### 2.2.2
 * Add ability to manage season buffs through side menu.
+* New filters: Multiple hits, Revive, Counter, Ignore defense
+* Improve verbiage for Haste, Revive and Counter effects
+* Fixes for PSY Nami GP condition and burst
 
 ### 2.2.1
 * Grand Party team builder.